+++ src/BabylonCpp/include/babylon/sprites/sprite.h
@@ -286,7 +286,7 @@ public:
   /**
    * Gets or sets the unique id of the sprite
    */
-  unsigned int uniqueId;
+  size_t uniqueId;
 
   /**
    * Gets the manager of this sprite
